test: extract inline mocks and define matchMedia mock
