+++ projects/client/test/mocks/matchMedia.mock.ts
@@ -1,0 +1,10 @@
+(globalThis as Record<string, unknown>).matchMedia = function () {
+  return {
+    matches: false,
+    media: '',
+    onchange: null,
+    addEventListener: function () {},
+    removeEventListener: function () {},
+    dispatchEvent: function () {},
+  };
+};
